Extract abstracts from Pubmed
Changed source code provided by instructor to extract the paper abstractions for project and put into text file.
