+++ KDM-ICP3/README.md
@@ -1,0 +1,23 @@
+# KDM-ICP3
+
+CS5560 Knowledge Discovery Management - In Class Programming 3
+
+# Introduction
+
+Before starting this ICP, the abstracts for the 10 papers needed to be extracted. The source code provided by the instructor was modified by me for my use case for the extraction.
+
+## Abstraction Steps:
+
+- Retrieve Abstracts using HttpURLConnection to Pubmed for an xml file.
+
+  - Use the [efetch xml example](https://dataguide.nlm.nih.gov/eutilities/utilities.html#efetch) from Pubmed and alter to get the abstract.
+
+  ![Abstract_extraction_xml](../docs/KDM-ICP3/Abstract_extraction_xml.gif)
+
+- From the XML file, put the abstracts into a text file for easier readability.
+
+  ![Abstract_xml_to_txt](../docs/KDM-ICP3/Abstract_xml_to_txt.gif)
+
+# Source Code
+
+The source code for this ICP was provided by the class instructor Mayanka ChandraShekar: [mckw9@mail.umkc.edu](mckw9@mail.umkc.edu)
