+++ KDM-ICP3/Retrieve_abstract/README.txt
@@ -1,10 +1,10 @@
-Retrive: https://www.dropbox.com/s/xpenw0o829b3a4t/Retrive.java?dl=0
+Retreive: https://www.dropbox.com/s/xpenw0o829b3a4t/Retrive.java?dl=0
 
-Retrive max of 100000 Paper IDs from PubMed. 
+Retreive max of 100000 Paper IDs from PubMed. 
 Using the REST API : http://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=pubmed&term=alzheimer%20disease&retmax=100000
 
 
-RetriveAbstracts : https://www.dropbox.com/s/c0i2ygyvm5zfh0s/RetrieveAbstracts.java?dl=0
+RetreiveAbstracts : https://www.dropbox.com/s/c0i2ygyvm5zfh0s/RetrieveAbstracts.java?dl=0
 
 Will retrieve 200 abstracts using IDS downloaded from previous rest call.
